refactor: move arc-0300 modal content to separate compnents
