+++ src/extension/components/ARC0300AccountImportWithAddressModalContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ARC0300AccountImportWithAddressModalContent';
