+++ src/extension/components/ARC0300AccountImportWithPrivateKeyModalContent/ARC0300AccountImportWithPrivateKeyModalContent.tsx
@@ -8,7 +8,7 @@ import {
   Text,
   VStack,
 } from '@chakra-ui/react';
-import React, { FC, useState } from 'react';
+import React, { FC, KeyboardEvent, useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { IoArrowBackOutline } from 'react-icons/io5';
 import { useDispatch } from 'react-redux';
@@ -23,6 +23,9 @@ import ModalSkeletonItem from '@extension/components/ModalSkeletonItem';
 import ModalItem from '@extension/components/ModalItem';
 import ModalTextItem from '@extension/components/ModalTextItem';
 import ModalSubHeading from '@extension/components/ModalSubHeading';
+import PasswordInput, {
+  usePassword,
+} from '@extension/components/PasswordInput';
 
 // constants
 import {
@@ -32,14 +35,18 @@ import {
 } from '@extension/constants';
 
 // enums
-import { AccountTabEnum, ARC0300QueryEnum } from '@extension/enums';
+import {
+  AccountTabEnum,
+  ARC0300QueryEnum,
+  ErrorCodeEnum,
+} from '@extension/enums';
 
 // features
 import {
   addARC0200AssetHoldingsThunk,
   IUpdateAssetHoldingsResult,
   saveActiveAccountDetails,
-  saveNewWatchAccountThunk,
+  saveNewAccountThunk,
 } from '@extension/features/accounts';
 import { create as createNotification } from '@extension/features/notifications';
 
@@ -53,7 +60,9 @@ import useUpdateARC0200Assets from '@extension/hooks/useUpdateARC0200Assets';
 import {
   useSelectActiveAccountDetails,
   useSelectLogger,
+  useSelectPasswordLockPassword,
   useSelectSelectedNetwork,
+  useSelectSettings,
 } from '@extension/selectors';
 
 // services
@@ -66,28 +75,32 @@ import { theme } from '@extension/theme';
 // types
 import type {
   IAccount,
-  IAppThunkDispatch,
   IARC0300AccountImportSchema,
-  IARC0300AccountImportWithAddressQuery,
+  IARC0300AccountImportWithPrivateKeyQuery,
+  IARC0300ModalContentProps,
+  IAppThunkDispatch,
 } from '@extension/types';
-import type { IModalContentProps } from './types';
 
 // utils
+import convertPrivateKeyToAddress from '@extension/utils/convertPrivateKeyToAddress';
 import ellipseAddress from '@extension/utils/ellipseAddress';
-import WatchAccountBadge from '@extension/components/WatchAccountBadge';
+import decodePrivateKeyFromAccountImportSchema from '@extension/utils/decodePrivateKeyFromImportKeySchema';
 
-const AccountImportWithAddressModalContent: FC<
-  IModalContentProps<
-    IARC0300AccountImportSchema<IARC0300AccountImportWithAddressQuery>
+const ARC0300AccountImportWithPrivateKeyModalContent: FC<
+  IARC0300ModalContentProps<
+    IARC0300AccountImportSchema<IARC0300AccountImportWithPrivateKeyQuery>
   >
 > = ({ onComplete, onPreviousClick, schema }) => {
   const { t } = useTranslation();
   const dispatch = useDispatch<IAppThunkDispatch>();
   const navigate = useNavigate();
+  const passwordInputRef = useRef<HTMLInputElement | null>(null);
   // selectors
   const activeAccountDetails = useSelectActiveAccountDetails();
   const logger = useSelectLogger();
   const network = useSelectSelectedNetwork();
+  const passwordLockPassword = useSelectPasswordLockPassword();
+  const settings = useSelectSettings();
   // hooks
   const defaultTextColor = useDefaultTextColor();
   const {
@@ -95,9 +108,18 @@ const AccountImportWithAddressModalContent: FC<
     loading,
     reset: resetUpdateAssets,
   } = useUpdateARC0200Assets(schema.query[ARC0300QueryEnum.Asset]);
+  const {
+    error: passwordError,
+    onChange: onPasswordChange,
+    reset: resetPassword,
+    setError: setPasswordError,
+    validate: validatePassword,
+    value: password,
+  } = usePassword();
   const primaryButtonTextColor = usePrimaryButtonTextColor();
   const subTextColor = useSubTextColor();
   // states
+  const [address, setAddress] = useState<string | null>(null);
   const [saving, setSaving] = useState<boolean>(false);
   // handlers
   const handlePreviousClick = () => {
@@ -105,17 +127,83 @@ const AccountImportWithAddressModalContent: FC<
     onPreviousClick();
   };
   const handleImportClick = async () => {
+    const _functionName: string = 'handleImportClick';
+    let _password: string | null;
     let account: IAccount | null;
     let actionTrackingService: ActionTrackingService;
+    let privateKey: Uint8Array | null;
     let result: IUpdateAssetHoldingsResult;
+
+    // if there is no password lock
+    if (!settings.security.enablePasswordLock && !passwordLockPassword) {
+      // validate the password input
+      if (validatePassword()) {
+        logger.debug(
+          `${ARC0300AccountImportWithPrivateKeyModalContent.name}#${_functionName}: password not valid`
+        );
+
+        return;
+      }
+    }
+
+    _password = settings.security.enablePasswordLock
+      ? passwordLockPassword
+      : password;
+
+    if (!_password) {
+      logger.debug(
+        `${ARC0300AccountImportWithPrivateKeyModalContent.name}#${_functionName}: unable to use password from password lock, value is "null"`
+      );
+
+      dispatch(
+        createNotification({
+          description: t<string>('errors.descriptions.code', {
+            context: ErrorCodeEnum.ParsingError,
+            type: 'password',
+          }),
+          ephemeral: true,
+          title: t<string>('errors.titles.code', {
+            context: ErrorCodeEnum.ParsingError,
+          }),
+          type: 'error',
+        })
+      );
+
+      return;
+    }
+
+    privateKey = decodePrivateKeyFromAccountImportSchema(schema);
+
+    if (!privateKey) {
+      logger.debug(
+        `${ARC0300AccountImportWithPrivateKeyModalContent.name}#${_functionName}: failed to decode the private key`
+      );
+
+      dispatch(
+        createNotification({
+          description: t<string>('errors.descriptions.code', {
+            context: ErrorCodeEnum.ParsingError,
+            type: 'key',
+          }),
+          ephemeral: true,
+          title: t<string>('errors.titles.code', {
+            context: ErrorCodeEnum.ParsingError,
+          }),
+          type: 'error',
+        })
+      );
+
+      return;
+    }
 
     setSaving(true);
 
     try {
       account = await dispatch(
-        saveNewWatchAccountThunk({
-          address: schema.query[ARC0300QueryEnum.Address],
+        saveNewAccountThunk({
           name: null,
+          password: _password,
+          privateKey,
         })
       ).unwrap();
 
@@ -133,6 +221,19 @@ const AccountImportWithAddressModalContent: FC<
       }
     } catch (error) {
       switch (error.code) {
+        case ErrorCodeEnum.InvalidPasswordError:
+          setPasswordError(t<string>('errors.inputs.invalidPassword'));
+
+          break;
+        case ErrorCodeEnum.PrivateKeyAlreadyExistsError:
+          logger.debug(
+            `${ARC0300AccountImportWithPrivateKeyModalContent.name}#${_functionName}: account already exists, carry on`
+          );
+
+          // clean up and close
+          handleOnComplete();
+
+          break;
         default:
           dispatch(
             createNotification({
@@ -191,14 +292,36 @@ const AccountImportWithAddressModalContent: FC<
     // clean up and close
     handleOnComplete();
   };
+  const handleKeyUpPasswordInput = async (
+    event: KeyboardEvent<HTMLInputElement>
+  ) => {
+    if (event.key === 'Enter') {
+      await handleImportClick();
+    }
+  };
   const handleOnComplete = () => {
     reset();
     onComplete();
   };
   const reset = () => {
+    resetPassword();
     resetUpdateAssets();
     setSaving(false);
   };
+
+  useEffect(() => {
+    if (passwordInputRef.current) {
+      passwordInputRef.current.focus();
+    }
+  }, []);
+  useEffect(() => {
+    const privateKey: Uint8Array | null =
+      decodePrivateKeyFromAccountImportSchema(schema);
+
+    if (privateKey) {
+      setAddress(convertPrivateKeyToAddress(privateKey));
+    }
+  }, []);
 
   return (
     <ModalContent
@@ -224,17 +347,18 @@ const AccountImportWithAddressModalContent: FC<
             <ModalSubHeading text={t<string>('labels.account')} />
 
             {/*address*/}
-            <ModalTextItem
-              label={`${t<string>('labels.address')}:`}
-              tooltipLabel={schema.query[ARC0300QueryEnum.Address]}
-              value={ellipseAddress(schema.query[ARC0300QueryEnum.Address], {
-                end: 10,
-                start: 10,
-              })}
-            />
-
-            {/*type*/}
-            <ModalItem label={t('labels.type')} value={<WatchAccountBadge />} />
+            {!address ? (
+              <ModalSkeletonItem />
+            ) : (
+              <ModalTextItem
+                label={`${t<string>('labels.address')}:`}
+                tooltipLabel={address}
+                value={ellipseAddress(address, {
+                  end: 10,
+                  start: 10,
+                })}
+              />
+            )}
           </VStack>
 
           {/*assets*/}
@@ -290,6 +414,17 @@ const AccountImportWithAddressModalContent: FC<
       {/*footer*/}
       <ModalFooter p={DEFAULT_GAP}>
         <VStack alignItems="flex-start" spacing={DEFAULT_GAP - 2} w="full">
+          {!settings.security.enablePasswordLock && !passwordLockPassword && (
+            <PasswordInput
+              error={passwordError}
+              hint={t<string>('captions.mustEnterPasswordToImportAccount')}
+              onChange={onPasswordChange}
+              onKeyUp={handleKeyUpPasswordInput}
+              inputRef={passwordInputRef}
+              value={password}
+            />
+          )}
+
           <HStack spacing={DEFAULT_GAP - 2} w="full">
             {/*previous button*/}
             <Button
@@ -319,4 +454,4 @@ const AccountImportWithAddressModalContent: FC<
   );
 };
 
-export default AccountImportWithAddressModalContent;
+export default ARC0300AccountImportWithPrivateKeyModalContent;
