+++ src/extension/components/ARC0300AccountImportWithPrivateKeyModalContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ARC0300AccountImportWithPrivateKeyModalContent';
