+++ src/extension/components/KeyRegistrationTransactionModalBody/index.ts
@@ -1,0 +1,2 @@
+export { default } from './KeyRegistrationTransactionModalBody';
+export * from './types';
