+++ src/extension/components/KeyRegistrationTransactionModalBody/types/IProps.ts
