+++ src/extension/components/KeyRegistrationTransactionModalBody/types/index.ts
