--- src/extension/components/KeyRegistrationTransactionModalContent/index.ts
@@ -1,2 +1,0 @@
-export { default } from './KeyRegistrationTransactionContent';
-export * from './types';
