+++ src/extension/enums/EventTypeEnum.ts
@@ -1,6 +1,6 @@
 enum EventTypeEnum {
+  ARC0300KeyRegistrationTransactionSend = 'event:arc0300_key_registration_transaction_send',
   ClientRequest = 'event:client_request',
-  SendKeyRegistrationTransaction = 'event:send_key_registration_transaction',
 }
 
 export default EventTypeEnum;
