+++ src/extension/events/index.ts
@@ -1,3 +1,3 @@
+export { default as ARC0300KeyRegistrationTransactionSendEvent } from './ARC0300KeyRegistrationTransactionSendEvent';
 export { default as ClientRequestEvent } from './ClientRequestEvent';
 export { default as BaseEvent } from './BaseEvent';
-export { default as SendKeyRegistrationTransactionEvent } from './SendKeyRegistrationTransactionEvent';
