+++ src/extension/features/events/thunks/handleNewEventByIdThunk.ts
@@ -32,7 +32,7 @@ const handleNewEventByIdThunk: AsyncThunk<
 
     switch (event.type) {
       case EventTypeEnum.ClientRequest:
-      case EventTypeEnum.SendKeyRegistrationTransaction:
+      case EventTypeEnum.ARC0300KeyRegistrationTransactionSend:
         return event;
       default:
         logger.debug(
