--- src/extension/modals/ScanQRCodeModal/types/index.ts
@@ -1,1 +1,0 @@
-export type { default as IModalContentProps } from './IModalContentProps';
