--- src/extension/types/events/ISendKeyRegistrationTransactionEventPayload.ts
@@ -1,19 +1,0 @@
-interface ISendKeyRegistrationTransactionEventPayload {
-  fee?: string;
-  firstValidRound?: string;
-  genesisHash?: string;
-  group?: string;
-  lastValidRound?: string;
-  lease?: string;
-  note?: string;
-  rekeyTo?: string;
-  sender: string;
-  selectionKey?: string;
-  stateProofKey?: string;
-  voteFirstRound?: string;
-  voteKey?: string;
-  voteKeyDilution?: string;
-  voteLastRound?: string;
-}
-
-export default ISendKeyRegistrationTransactionEventPayload;
