+++ src/extension/types/events/TEvents.ts
@@ -2,10 +2,10 @@ import { TRequestParams } from '@agoralabs-sh/avm-web-provider';
 
 // types
 import type IClientRequestEvent from './IClientRequestEvent';
-import type ISendKeyRegistrationTransactionEvent from './ISendKeyRegistrationTransactionEvent';
+import type IARC0300KeyRegistrationTransactionSendEvent from './IARC0300KeyRegistrationTransactionSendEvent';
 
 type TEvents =
-  | IClientRequestEvent<TRequestParams>
-  | ISendKeyRegistrationTransactionEvent;
+  | IARC0300KeyRegistrationTransactionSendEvent
+  | IClientRequestEvent<TRequestParams>;
 
 export default TEvents;
