+++ src/extension/types/events/index.ts
@@ -1,6 +1,5 @@
+export type { default as IARC0300KeyRegistrationTransactionSendEvent } from './IARC0300KeyRegistrationTransactionSendEvent';
 export type { default as IBaseEvent } from './IBaseEvent';
 export type { default as IClientRequestEvent } from './IClientRequestEvent';
 export type { default as IClientRequestEventPayload } from './IClientRequestEventPayload';
-export type { default as ISendKeyRegistrationTransactionEvent } from './ISendKeyRegistrationTransactionEvent';
-export type { default as ISendKeyRegistrationTransactionEventPayload } from './ISendKeyRegistrationTransactionEventPayload';
 export type { default as TEvents } from './TEvents';
