+++ src/extension/types/index.ts
@@ -53,13 +53,11 @@ export type { default as IAssetTransferTransaction } from './IAssetTransferTrans
 export type { default as IAssetTypes } from './IAssetTypes';
 export type { default as IAssetUnfreezeTransaction } from './IAssetUnfreezeTransaction';
 export type { default as IAsyncThunkConfigWithRejectValue } from './IAsyncThunkConfigWithRejectValue';
-export type { default as IBackgroundRootState } from './IBackgroundRootState';
 export type { default as IBaseActionMeta } from './IBaseActionMeta';
 export type { default as IBaseAsset } from './IBaseAsset';
 export type { default as IBaseAssetFreezeTransaction } from './IBaseAssetFreezeTransaction';
 export type { default as IBaseAssetHolding } from './IBaseAssetHolding';
 export type { default as IBaseAsyncThunkConfig } from './IBaseAsyncThunkConfig';
-export type { default as IBaseRootState } from './IBaseRootState';
 export type { default as IBaseTransaction } from './IBaseTransaction';
 export type { default as IBlockExplorer } from './IBlockExplorer';
 export type { default as IChainNamespace } from './IChainNamespace';
@@ -73,8 +71,6 @@ export type { default as IGeneralSettings } from './IGeneralSettings';
 export type { default as IInitializeAccountOptions } from './IInitializeAccountOptions';
 export type { default as IKeyRegistrationOfflineTransaction } from './IKeyRegistrationOfflineTransaction';
 export type { default as IKeyRegistrationOnlineTransaction } from './IKeyRegistrationOnlineTransaction';
-export type { default as IMainRootState } from './IMainRootState';
-export type { default as IModalProps } from './IModalProps';
 export type { default as INativeCurrency } from './INativeCurrency';
 export type { default as INetwork } from './INetwork';
 export type { default as INetworkWithTransactionParams } from './INetworkWithTransactionParams';
@@ -89,7 +85,6 @@ export type { default as IPendingActionMeta } from './IPendingActionMeta';
 export type { default as IPrivacySettings } from './IPrivacySettings';
 export type { default as IPrivateKey } from './IPrivateKey';
 export type { default as IRegistrationAddAccountCompleteResult } from './IRegistrationAddAccountCompleteResult';
-export type { default as IRegistrationRootState } from './IRegistrationRootState';
 export type { default as IRejectedActionMeta } from './IRejectedActionMeta';
 export type { default as IResourceLanguage } from './IResourceLanguage';
 export type { default as IScanQRCodeModalContentProps } from './IScanQRCodeModalContentProps';
@@ -108,3 +103,5 @@ export type { default as IVestigeFiAssetResponse } from './IVestigeFiAssetRespon
 export type { default as IWalletConnectSessionMetadata } from './IWalletConnectSessionMetadata';
 export * from './arc-0300';
 export * from './events';
+export * from './modals';
+export * from './states';
