+++ src/extension/types/modals/IARC0300ModalContentProps.ts
@@ -1,9 +1,9 @@
 import type { IARC0300BaseSchema } from '@extension/types';
 
-interface IModalContentProps<Schema extends IARC0300BaseSchema> {
+interface IARC0300ModalContentProps<Schema extends IARC0300BaseSchema> {
   onComplete: () => void;
   onPreviousClick: () => void;
   schema: Schema;
 }
 
-export default IModalContentProps;
+export default IARC0300ModalContentProps;
