+++ src/extension/types/modals/index.ts
@@ -1,0 +1,2 @@
+export type { default as IARC0300ModalContentProps } from './IARC0300ModalContentProps';
+export type { default as IModalProps } from './IModalProps';
