+++ src/extension/types/states/IBackgroundRootState.ts
