+++ src/extension/types/states/index.ts
@@ -1,0 +1,4 @@
+export type { default as IBackgroundRootState } from './IBackgroundRootState';
+export type { default as IBaseRootState } from './IBaseRootState';
+export type { default as IMainRootState } from './IMainRootState';
+export type { default as IRegistrationRootState } from './IRegistrationRootState';
